+++ Sources/RaiderIO/Guild/BossKill.swift
@@ -11,21 +11,23 @@ public struct BossKill {
 
     public struct Kill {
 
-        public let pulledAt: ISO8601Date
+        public let pulledAt: ISO8601Date?
         public let defeatedAt: ISO8601Date
-        public let duration: Milliseconds
+        public let duration: Milliseconds?
         public let isSuccess: Bool
-        public let itemLevelEquippedAverage: Float
-        public let itemLevelEquippedMax: Float
-        public let itemLevelEquippedMin: Float
+        public let itemLevelEquippedAverage: Float?
+        public let itemLevelEquippedMax: Float?
+        public let itemLevelEquippedMin: Float?
+        public let artifactTraitsAverage: Double?
 
-        public init(pulledAt: ISO8601Date,
+        public init(pulledAt: ISO8601Date?,
                     defeatedAt: ISO8601Date,
-                    duration: Milliseconds,
+                    duration: Milliseconds?,
                     isSuccess: Bool,
-                    itemLevelEquippedAverage: Float,
-                    itemLevelEquippedMax: Float,
-                    itemLevelEquippedMin: Float) {
+                    itemLevelEquippedAverage: Float?,
+                    itemLevelEquippedMax: Float?,
+                    itemLevelEquippedMin: Float?,
+                    artifactTraitsAverage: Double?) {
             self.pulledAt = pulledAt
             self.defeatedAt = defeatedAt
             self.duration = duration
@@ -33,6 +35,7 @@ public struct BossKill {
             self.itemLevelEquippedAverage = itemLevelEquippedAverage
             self.itemLevelEquippedMax = itemLevelEquippedMax
             self.itemLevelEquippedMin = itemLevelEquippedMin
+            self.artifactTraitsAverage = artifactTraitsAverage
         }
 
     }
@@ -62,6 +65,7 @@ extension BossKill.Kill: Codable {
         case itemLevelEquippedAverage   = "itemLevelEquippedAvg"
         case itemLevelEquippedMax
         case itemLevelEquippedMin
+        case artifactTraitsAverage      = "artifactTraitsAvg"
 
     }